# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "clipos API",
    "description": "Invite-only clip archive.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tags": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Tag autocomplete, most used first.",
        "operationId": "search_tags",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Prefix to complete.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Every active member, by name.",
        "operationId": "list_members",
        "responses": {
          "200": {
            "description": "Members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Member"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/{handle}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "A member's profile. Their clips: `GET /api/clips?uploader={handle}`.",
        "operationId": "get_profile",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "description": "Handle",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Error payload for every non-2xx `/api` response.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code: `bad_request`, `unauthorized`, `not_invited`,\n`account_disabled`, `forbidden`, `not_found`, `conflict`, `unsupported_media_type`,\n`too_large`, `bad_gateway` or `internal`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Never contains internals for 5xx responses."
          },
          "tied": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Ties",
                "description": "Only on the 409 from ending a show with a tied vote: the tied clips of every tied\ncategory, so the host breaks them all at once."
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "Member": {
        "type": "object",
        "description": "A user as other members see them (no email).",
        "required": [
          "id",
          "handle",
          "displayName"
        ],
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "steamName": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Profile": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Member"
          },
          {
            "type": "object",
            "required": [
              "clipCount",
              "featuredCount",
              "fireCount",
              "joinedAt"
            ],
            "properties": {
              "clipCount": {
                "type": "integer",
                "format": "int64",
                "description": "Clips they uploaded that you see: as many as `GET /api/clips?uploader={handle}`\nlists for you. Others' clips saved for the show (decision 29) and unpublished\nclips count only for their uploader."
              },
              "featuredCount": {
                "type": "integer",
                "format": "int64",
                "description": "Clips they play in (tagged by a friend), as many as `?player={handle}` lists."
              },
              "fireCount": {
                "type": "integer",
                "format": "int64",
                "description": "🔥 reactions their published clips got."
              },
              "joinedAt": {
                "type": "string",
                "format": "date-time",
                "description": "When they first signed in."
              }
            }
          }
        ]
      },
...
      "Ties": {
        "type": "object",
        "description": "The tied clips of each category when `end` needs the host's pick; empty for a\ncategory that isn't tied.",
        "required": [
          "clip",
          "fail"
        ],
        "properties": {
          "clip": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "fail": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}